+++ src/lib/components/FileUploader/TabUpload/TabUpload.svelte
@@ -221,7 +221,7 @@
 						block
 						placeholder="Enter image URL"
 						bind:value={byUrl}
-						on:keyup={(e) => e.key === 'Enter' && handleFetch()}
+						onkeyup={(e) => e.key === 'Enter' && handleFetch()}
 						bind:input={byUrlInputEl}
 					/>
 					<Button disabled={byUrl.trim() === ''} on:click={handleFetch}>